requirements.txt
argparse==1.4.0
configparser==4.0.2
dnspython==1.12.0
iniparse==0.3.1
iotop==0.3.2
ordereddict==1.2
pycurl==7.19.0
pygpgme==0.1
pyzor==1.0.0
urlgrabber==3.9.1
yum-metadata-parser==1.1.2
